Short scalar cast

+++ litespeed-cache/lib/litespeed/litespeed-string.class.php
@@ -19,7 +19,7 @@ class Litespeed_String
 	 */
 	public static function rrand( $len, $type = 7 )
 	{
-		mt_srand( ( double ) microtime() * 1000000 ) ;
+		mt_srand( ( float ) microtime() * 1000000 ) ;
 
 		switch( $type ) {
 			case 0 :
